Seems to be compiling, except for some undefined reference errors. Eigenified a heap of stuff (probably poorly).

+++ module/motion/OldWalkEngine/src/OldWalkEngine.h
@@ -165,7 +165,7 @@ namespace motion {
         double phase1Single;
         double phase2Single;
         Eigen::Vector2d footOffset;
-        double legYaw;
+        float legYaw;
         // Ankle feedback parameters
         double ankleRollComp;
         double ankleRollLimit;
